:art: add borax.strings
