+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "borax"
-version = "1.3.0"
+version = "1.4.0"
 description = "A util collections for Python3."
 readme = "long_description.rst"
 authors = ["kinegratii <kinegratii@gmail.com>"]
